rename the keyword/text fields
