+++ prismic-model/src/event-series.ts
@@ -3,7 +3,7 @@ import promo from './parts/promo';
 import body from './parts/body';
 import contributorsWithTitle from './parts/contributorsWithTitle';
 import { documentLink } from './parts/link';
-import { singleLineText } from './parts/structured-text';
+import { singleLineText } from './parts/text';
 import { CustomType } from './types/CustomType';
 
 const eventSeries: CustomType = {
