+++ prismic-model/src/global-alert.ts
@@ -1,6 +1,6 @@
-import { multiLineText } from './parts/structured-text';
+import { multiLineText } from './parts/text';
 import select from './parts/select';
-import text from './parts/text';
+import keyword from './parts/keyword';
 import { CustomType } from './types/CustomType';
 
 const globalAlert: CustomType = {
@@ -19,7 +19,7 @@ const globalAlert: CustomType = {
         defaultValue: 'hide',
         placeholder: 'Show or hide',
       }),
-      routeRegex: text(
+      routeRegex: keyword(
         'Write a pipe-separated (|) list of page paths here if you only want the alert to display on certain pages. Leave empty if you want the alert to appear on all pages.',
         { placeholder: 'path(s) to match' }
       ),
