+++ prismic-model/src/parts/article-body.ts
@@ -1,8 +1,8 @@
 import body, { slice } from './body';
 import heading from './heading';
 import { mediaLink, webLink } from './link';
-import text from './text';
-import { multiLineText, singleLineText } from './structured-text';
+import text from './keyword';
+import { multiLineText, singleLineText } from './text';
 import gifVideoSlice from './gif-video-slice';
 import title from './title';
