+++ prismic-model/src/parts/body.ts
@@ -1,4 +1,4 @@
-import { multiLineText, singleLineText } from './structured-text';
+import { multiLineText, singleLineText } from './text';
 import captionedImageSlice from './captioned-image-slice';
 import captionedImageGallerySlice from './captioned-image-gallery-slice';
 import gifVideoSlice from './gif-video-slice';
@@ -6,7 +6,7 @@ import iframeSlice from './iframe-slice';
 import table from './table';
 import title from './title';
 import link, { documentLink, mediaLink, webLink } from './link';
-import text from './text';
+import text from './keyword';
 import embed from './embed';
 import mediaObject from './media-object';
 import heading from './heading';
