+++ prismic-model/src/parts/description.ts
@@ -1,4 +1,4 @@
-import { multiLineText } from './structured-text';
+import { multiLineText } from './text';
 
 const description = multiLineText('Description', {
   extraTextOptions: ['heading2'],
