+++ prismic-model/src/parts/text.ts
@@ -1,13 +1,61 @@
-type Props = { placeholder: string };
+type TextProps = {
+  extraTextOptions?: string[];
+  overrideTextOptions?: string[];
+  placeholder?: string;
+};
 
-// This should be used sparingly as you can't free text search on it, only
-// exact match
-export default function (label: string, props?: Props) {
+type StructuredTextProps = TextProps & {
+  label: string;
+  allowMultipleParagraphs?: boolean;
+};
+
+const defaultTextOptions = ['paragraph', 'hyperlink', 'strong', 'em'];
+
+function structuredText({
+  label,
+  allowMultipleParagraphs = true,
+  extraTextOptions = [],
+  overrideTextOptions = [],
+  placeholder,
+}: StructuredTextProps) {
+  // See https://prismic.io/docs/technologies/rich-text-title#json-model
+  const singleOrMulti = allowMultipleParagraphs ? 'multi' : 'single';
+
+  if (overrideTextOptions.length > 0 && extraTextOptions.length > 0) {
+    throw new Error(
+      'Either specify extra text options or a complete list; not both'
+    );
+  }
+
+  const textOptions =
+    overrideTextOptions.length > 0
+      ? overrideTextOptions.join(',')
+      : [...defaultTextOptions, ...extraTextOptions].join(',');
+
   return {
-    type: 'Text',
+    type: 'StructuredText',
     config: {
-      label,
-      placeholder: props?.placeholder,
+      // This is too complicated but it's because we've overloaded this type with things like keywords.
+      // See interpretation-types.abbreviation for an example.
+      [singleOrMulti]: textOptions,
+      label: label,
+      placeholder,
     },
   };
+}
+
+export function singleLineText(label: string, props?: TextProps) {
+  return structuredText({
+    label,
+    ...props,
+    allowMultipleParagraphs: false,
+  });
+}
+
+export function multiLineText(label: string, props?: TextProps) {
+  return structuredText({
+    label,
+    ...props,
+    allowMultipleParagraphs: true,
+  });
 }
