+++ prismic-model/src/people.ts
@@ -1,8 +1,8 @@
-import text from './parts/text';
+import keyword from './parts/keyword';
 import description from './parts/description';
 import image from './parts/image';
 import list from './parts/list';
-import { singleLineText } from './parts/structured-text';
+import { singleLineText } from './parts/text';
 import { CustomType } from './types/CustomType';
 
 const people: CustomType = {
@@ -12,12 +12,12 @@ const people: CustomType = {
   status: true,
   json: {
     Person: {
-      name: text('Full name'),
+      name: keyword('Full name'),
       description: description,
-      pronouns: text('Pronouns'),
+      pronouns: keyword('Pronouns'),
       image: image('Image'),
       sameAs: list('Same as', {
-        link: text('Link'),
+        link: keyword('Link'),
         title: singleLineText('Link text'),
       }),
     },
