+++ prismic-model/src/places.ts
@@ -2,7 +2,7 @@ import title from './parts/title';
 import geolocation from './parts/geolocation';
 import number from './parts/number';
 import body from './parts/body';
-import { multiLineText } from './parts/structured-text';
+import { multiLineText } from './parts/text';
 import { CustomType } from './types/CustomType';
 
 const places: CustomType = {
