+++ prismic-model/src/popup-dialog.ts
@@ -1,5 +1,5 @@
-import { singleLineText } from './parts/structured-text';
-import text from './parts/text';
+import { singleLineText } from './parts/text';
+import keyword from './parts/keyword';
 import boolean from './parts/boolean';
 import { webLink } from './parts/link';
 import { CustomType } from './types/CustomType';
@@ -11,10 +11,10 @@ const popupDialog: CustomType = {
   status: true,
   json: {
     'Popup dialog': {
-      openButtonText: text('Open button text'),
-      title: text('Title inside the open dialog'),
+      openButtonText: keyword('Open button text'),
+      title: keyword('Title inside the open dialog'),
       text: singleLineText('Text inside the open dialog'),
-      linkText: text('CTA inside the open dialog button text'),
+      linkText: keyword('CTA inside the open dialog button text'),
       link: webLink('CTA inside the open dialog button link'),
       isShown: boolean('Is shown?', { defaultValue: false }),
     },
