+++ prismic-model/src/projects.ts
@@ -1,7 +1,7 @@
 import title from './parts/title';
 import body from './parts/body';
 import promo from './parts/promo';
-import { singleLineText } from './parts/structured-text';
+import { singleLineText } from './parts/text';
 import contributorsWithTitle from './parts/contributorsWithTitle';
 import { documentLink } from './parts/link';
 import list from './parts/list';
