+++ prismic-model/src/types/label.ts
@@ -1,5 +1,5 @@
 import title from '../parts/title';
-import { multiLineText } from '../parts/structured-text';
+import { multiLineText } from '../parts/text';
 
 export default function label(label: string) {
   return {
