+++ prismic-model/src/webcomic-series.ts
@@ -1,7 +1,7 @@
 // This has been deprecated and we should be using the article format of comic
 // and general article series.
 import title from './parts/title';
-import { multiLineText, singleLineText } from './parts/structured-text';
+import { multiLineText, singleLineText } from './parts/text';
 import promo from './parts/promo';
 import contributorsWithTitle from './parts/contributorsWithTitle';
 import { CustomType } from './types/CustomType';
